We had several modules with the same name "Search". Trying to avoid that.
