+++ src/js/WebParts/ExperienceLog/ExperienceLogSearch.ts
@@ -37,7 +37,7 @@ export interface IQueryResponse {
 /**
  * Query the REST Search API using sp-pnp-js
  */
-export const query = (SelectProperties: string[]) => new Promise<IQueryResponse>((resolve, reject) => {
+export const queryLogElements = (SelectProperties: string[]) => new Promise<IQueryResponse>((resolve, reject) => {
     SelectProperties = SelectProperties.concat(["Path", "SPWebUrl"]);
     sp.search({
         ...DEFAULT_SEARCH_SETTINGS,
